requirements.txt
matplotlib 
opencv-python 
scikit-learn
seaborn
tqdm
torchinfo
tslearn
fastdtw
autopep8
hydra-core
accelerate == 0.24.1